migrated all client/src/components from jsx to tsx
