+++ www/client/package.json
@@ -57,6 +57,7 @@
     ]
   },
   "devDependencies": {
+    "@types/react-modal": "^3.16.3",
     "@types/react-color": "^3.0.13",
     "@types/react-slider": "^1.3.6",
     "@vitejs/plugin-react": "^5.0.1",
