+++ www/client/src/components/DemoButton.tsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import NavButton from './NavButton';
+import { backendUrl } from '../config';
+
+interface DemoButtonProps {
+  children?: React.ReactNode;
+  [key: string]: any;
+}
+
+const DemoButton: React.FC<DemoButtonProps> = (props) => {
+  return (
+    <NavButton {...props} action={() => {
+      const requestOptions = {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify({ compositionId: 0 })
+      };
+      fetch(backendUrl + '/play', requestOptions);
+    }}>
+      {props.children}
+    </NavButton>
+  );
+}
+
+export default DemoButton;
